fix(auth): fix auth api
